Introduced the Composition Root
In *ASP.NET Core*, the composition root is no longer an `IControllerFactory`, but is now an `IControllerActivator`. The logic within this is largely the same. This commit establishes a new `SampleControllerActivator`, hooks up the existing `HomeController` (not part of *OnTopic*`, and registers the activator as part of the `Startup` class.