Introduced `CreateTopicController()` to `IControllerActivator`
The original `IControllerActivator` only supported the OOTB `HomeController`. With the dependencies added, we can now additionally support the *OnTopic* `TopicController`. Note that since `HttpRequest` no longer returns a `Uri` object, its `Path` must be converted to a `Uri` first—this may suggest a need to refactor this in the future to better accommodate this (e.g., by accepting the path, instead of requiring a `Uri`).