Ensure that the view model's assembly is loaded (HACK)
This is a bit of a hack—and there are other ways of addressing it which we may explore in the future—but the `ITypeLookupService` won't be capable of finding any of the topic view models if the `Ignia.Topics.ViewModels` assembly isn't opened. And, in *ASP.NET Core*, assemblies are not opened or even referenced from the project unless they've been called. To mitigate this, an instance of the `PageTopicViewModel` is created—even though it'll never be used by the `SampleControllerActivator`.