Introduced models from *GoldSim*
For test purposes, the site is currently running against the *GoldSim* copy of the *OnTopic* database, since it's a complex, comprehensive, and mature implementation with a lot of different topics. As a result, however, it also has a lot of _custom_ content types. While not all of those are necessary for test purposes (e.g., `Application*`, `Example*`), others are necessary just to test basic functionality (e.g., `FAQ` in order to test the `Accordion` view).  These will likely need to be trimmed down further for a reference baseline. Until then, however, they enable deeper integration testing. The namespace has been changed to `OnTopicTest.Models` for consistency.