Introduced basic styles
Introduced basic styles based on the *GoldSim* templates. `FAQ` is specific to a *GoldSim* content type, but is introduced to demonstrate the `Accordion` view, which `FAQ` refers to. Again, these are using *Foundation* styles, so they're not going to display until either a) *Zurb* is introduced, or b) they are retrofitted to use *Bootstrap*.