Introduced `MenuViewComponent` concrete type
The `MenuViewComponentBase{T}` that ships with the `Ignia.Topics.AspNetCore.Mvc` library is generic, which is not supported for view components. To mitigate that, while still allowing it to remain view model agnostic, it must be derived by a concrete class that sets the genric type for the local application. This class needn't provide any further functionality.