Refactor C# syntax in Razor tag helper
In *ASP.NET Core 3*, the `<option />` element is natively understood as a tag helper. This provides a number of useful features, including smarter intellisense and data binding. But it also means we can have *C#* code embedded into it. As such, refactored this code to put the *C#* expression within an attribute, to comply with the *Razor* syntax expectations.